correcciones, y tabla DetalleCuotasPrestamos, con su lógica y DAO

+++ src/logic/SolicitudPrestamoLogic.java
@@ -168,7 +168,7 @@ public class SolicitudPrestamoLogic implements IRecordLogic<SolicitudPrestamo,St
 		s.setFechaPedido(d.$("fechaPedido_Sol"));
 		s.setMontoPedido(d.$("montoPedido_Sol"));
 		s.setMontoAPagar(d.$("montoAPagar_Sol"));
-		s.setInteres(d.$("interes_Sol"));
+		s.setInteres((float)d.$("interes_Sol"));
 		s.setPlazoPago(d.$("plazoPago_Sol"));
 		s.setCantCuotas(d.$("cantCuotas_Sol"));
 		s.setMontoPorCuota(d.$("montoPorCuota_Sol"));
@@ -188,4 +188,3 @@ public class SolicitudPrestamoLogic implements IRecordLogic<SolicitudPrestamo,St
 	}
 
 }
-
